modulo gestion de usuario creado sin validaciones de permisos

+++ internal/app/application/users/deleteUser.go
@@ -5,21 +5,30 @@ import (
 	"shipping-app/internal/app/domain/ports/repository"
 )
 
-type DeleteUser struct {
-	repo repository.UserRepository 
+var (
+	ErrUserNotFoundDelete = errors.New("Usuario no encontrado")
+)
+
+type DeleteUserUseCase struct {
+	repo repository.UserRepository
 }
 
-func NewDeleteUserUseCase(repo repository.UserRepository) *DeleteUser {
-	return &DeleteUser{repo: repo} 
+func NewDeleteUserUseCase(repo repository.UserRepository) *DeleteUserUseCase {
+	return &DeleteUserUseCase{repo: repo}
 }
 
-func (uc *DeleteUser) Execute(id uint) error { 
-	
-	_, err := uc.repo.GetUserByID(id) 
-	if err != nil {
-		return errors.New("usuario no encontrado")
+func (uc *DeleteUserUseCase) Execute(id uint) error {
+	// 1. Validar ID
+	if id == 0 {
+		return ErrInvalidID
 	}
 
-	
+	// 2. Verificar que el usuario existe
+	_, err := uc.repo.GetUserByID(id)
+	if err != nil {
+		return ErrUserNotFound
+	}
+
+	// 3. Eliminar usuario
 	return uc.repo.DeleteUser(id)
 }
